Replace subscriberasync listener by dyn version
