+++ dds/src/dds_async/subscriber_listener.rs
@@ -1,4 +1,4 @@
-use std::future::Future;
+use std::{future::Future, pin::Pin};
 
 use crate::{
     infrastructure::status::{
@@ -15,69 +15,90 @@ pub trait SubscriberListenerAsync {
     /// Method that is called when any reader belonging to this subcriber reports new data available. This method is triggered before the on_data_available method.
     fn on_data_on_readers(
         &mut self,
-        _the_subscriber: SubscriberAsync,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+        the_subscriber: SubscriberAsync,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + '_>> {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports new data available.
-    fn on_data_available(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_data_available<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
     /// Method that is called when any reader belonging to this subcriber reports a sample rejected status.
-    fn on_sample_rejected(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: SampleRejectedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_sample_rejected<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: SampleRejectedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports a liveliness changed status.
-    fn on_liveliness_changed(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: LivelinessChangedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_liveliness_changed<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: LivelinessChangedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports a requested deadline missed status.
-    fn on_requested_deadline_missed(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: RequestedDeadlineMissedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_requested_deadline_missed<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: RequestedDeadlineMissedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports a requested incompatible QoS status.
-    fn on_requested_incompatible_qos(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: RequestedIncompatibleQosStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_requested_incompatible_qos<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: RequestedIncompatibleQosStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports a subscription matched status.
-    fn on_subscription_matched(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: SubscriptionMatchedStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_subscription_matched<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: SubscriptionMatchedStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 
     /// Method that is called when any reader belonging to this subcriber reports a sample lost status.
-    fn on_sample_lost(
-        &mut self,
-        _the_reader: &dyn AnyDataReader,
-        _status: SampleLostStatus,
-    ) -> impl Future<Output = ()> + Send {
-        std::future::ready(())
+    fn on_sample_lost<'a, 'b>(
+        &'a mut self,
+        the_reader: &'b (dyn AnyDataReader + Sync),
+        status: SampleLostStatus,
+    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'b>>
+    where
+        'a: 'b,
+    {
+        Box::pin(std::future::ready(()))
     }
 }
